Yeah, so, we print final statistics now.
Added state variables to Employee, so each Employee can keep track of
its statistics.  Modified methods doWork(~),  busyWait(~), busyWait().
Modified for all classes overriding these methods as well (No classes
over-rode these methods).

Changed TeamLead's answerQuestion() so it will keep track of how long it
waits for the PM to answer a question.  Will multiply how long it waits
by 2 because of the developer that waits with him.  Also subtracts the
time it took the PM to answer the question, as this doesn't count as
time waiting to have a question answered.

Added a method to Firm to print out the stuff.

Added utility method to FirmTime called calculateDifference() because I
needed to figure out the difference between two times. Probably
should've found something in the JAVA API - but hey - it was lik 3 AM at
the time.

Working out a couple kinks, but figured I'd merge in the lion's share of
stuff.